requirements.txt
openai>=1.12.0
pydantic>=2.5.0
python-dotenv>=1.0.0
numpy>=1.24.0
tqdm>=4.66.0
